\name{bigV3-package}
\alias{bigV3-package}
\alias{bigV3}
\docType{package}
\title{
\packageTitle{bigV3}
}
\description{
\packageDescription{bigV3}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{bigV3}
\packageIndices{bigV3}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{bigV3}

Maintainer: \packageMaintainer{bigV3}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
